Incorporate refx as Redux side effect pattern
